feat: improved md preview with code and emojis
It should be refactored so it can be applied to the feed at large.
Currently, the existing markdown rendering for the Posts themselves are
being done in-server by Redcarpet, and it lacks the styling apparent in
the preview.

Not sure how to refactor the stimulus controller to run efficienctly to
reasonably apply that to the feed client-side.